loopy.ld: add LMA alignment statements
This exclusively affects output sections that use the `AT>`
keyword.

ld appears to have curious behavior where the alignment requirements
of a section only respected as a coincidence of the case where
LMA==VMA. If an alternate LMA is specified with AT> , alignment
requirements are completely ignored when the LMA is created.

In the case of loopy.ld, if (the last input section of) .rodata has a
size that is not an even multiple of 4, and the size of .ctors is
zero, the LMA chosen for .data will be an odd/unaligned address.

There does not appear to be a convenient way to convince the linker to
preserve the section's alignment requirements when an AT> LMA is
chosen.

In an attempt to try to generate correct LMAs, this adds the
statement:

  . = ALIGN(4)

to the end of every output section prior to .data. This causes each
output section to have a size that is a multiple of 4.

I verified that this indeed appears to prevent LMA misalignment for
output sections that use the AT> keyword.